# Optimize Rubyboy::Bus