Uses the same style as: https://github.com/sacckey/rubyboy/pull/24

main:

```
Ruby: 3.4.1
YJIT: true
1: 3.882817 sec
2: 3.83399 sec
3: 3.835251 sec
FPS: 389.5409804902295

Ruby: 3.4.1
YJIT: true
1: 3.981531 sec
2: 3.804446 sec
3: 3.823725 sec
FPS: 387.606848134431
```

This branch:

```
Ruby: 3.4.1
YJIT: true
1: 3.503513 sec
2: 3.437195 sec
3: 3.52252 sec
FPS: 430.07760129092094
```

```
Ruby: 3.4.1
YJIT: true
1: 3.620516 sec
2: 3.582359 sec
3: 3.484475 sec
FPS: 421.05854117250766
```