@byroot 2025-02-10
Redoing benchmarks after rebasing now that #24 was merged:

Before:
```
Ruby: 3.4.1
YJIT: true
1: 3.476809 sec
2: 3.597895 sec
3: 3.617027 sec
FPS: 420.8860099454429

Ruby: 3.4.1
YJIT: true
1: 3.606878 sec
2: 3.63947 sec
3: 3.637668 sec
FPS: 413.45032936372013
```

After:
```
Ruby: 3.4.1
YJIT: true
1: 3.362835 sec
2: 3.319436 sec
3: 3.426609 sec
FPS: 445.15317226042845

Ruby: 3.4.1
YJIT: true
1: 3.394805 sec
2: 3.419763 sec
3: 3.438782 sec
FPS: 438.8809511037856
```